Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -69,6 +69,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        /* Condition from AuthServiceProvider.php */
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -81,6 +85,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /* Condition from AuthServiceProvider.php */
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated.');
@@ -94,6 +103,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        /* Condition from AuthServiceProvider.php */
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
